+++ internal/provider/provider.go
@@ -522,6 +522,7 @@ func (p *MerakiProvider) DataSources(ctx context.Context) []func() datasource.Da
 		NewWirelessSSIDSplashSettingsDataSource,
 		NewWirelessSSIDTrafficShapingRulesDataSource,
 		NewWirelessSSIDVPNDataSource,
+		NewWirelessSSIDsDataSource,
 	}
 }